Remove force=true for JSON content types and refactor tests in to single sign_in tests
